+++ lib/constants/authentication_string.dart
@@ -1,0 +1,18 @@
+// ignore_for_file: constant_identifier_names
+
+class AuthenticationString {
+  static const MAINHEADING1 = 'Enter your';
+  static const MAINHEADING2 = 'mobile number';
+  static const TRANSPARENTTEXT = 'TRANSPARENT TEXT';
+  static const SUBHEADING = 'we will send you a confirmation code';
+  static const COUNTRYCODE = '+91  ';
+  static const NEXTBTN = 'Next';
+  static const NUMBERVALIDATION =
+      'Don\'t be dumb, number should be 10 digit long';
+  static const OTPVALIDATION = 'Otp is 6 Digit Long';
+  static const OTPHEADING1 = 'Enter ';
+  static const OTPHEADING2 = 'verification code';
+  static const OTPSUBHEADING = 'we have send you on +91 ';
+  static const OTPRESEND = 'Resend otp';
+  static const CONFIRMBTN = 'Confirm';
+}