+++ lib/constants/constants.dart
@@ -1,1 +1,2 @@
 export 'splash_string.dart';
+export 'authentication_string.dart';