+++ lib/utils/router/nav_links.dart
@@ -1,6 +1,7 @@
+// ignore_for_file: constant_identifier_names
+
 class NavLinks {
-  // ignore: constant_identifier_names
   static const SPLASH = "/splash";
-  // ignore: constant_identifier_names
+  static const PHONEAUTH = "/phoneAuth";
   static const HOME = "/home";
 }